% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GctMatrix.R
\name{as.matrix.GctMatrix}
\alias{as.matrix.GctMatrix}
\title{Coerce a GctMatrix object into a matrix}
\usage{
\method{as.matrix}{GctMatrix}(x, ...)
}
\arguments{
\item{x}{A GctMatrix object}
}
\value{
A matrix with a \code{desc} attribute
}
\description{
Coerce a GctMatrix object into a matrix
}
\examples{
m1 <- matrix(1:6, nrow=3, dimnames=list(sprintf("G\%d", 1:3), sprintf("S\%d", 1:2)))
gm1 <- GctMatrix(m1, desc=sprintf("Gene\%d", 1:3))
print(gm1)
print(as.matrix(gm1))
}
