% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exprs_matrix.R
\name{as_numeric_matrix}
\alias{as_numeric_matrix}
\title{Transform a data.frame to a numeric matrix without characters coereced as factors}
\usage{
as_numeric_matrix(df, warning = FALSE)
}
\arguments{
\item{df}{A data.frame}

\item{warning}{Logical, whether the function should warn when non-numeric
characters are transformed}
}
\value{
A numeric matrix
}
\description{
Transform a data.frame to a numeric matrix without characters coereced as factors
}
