% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{loadObject}
\alias{loadObject}
\title{Load an object by its name from a RData file}
\usage{
loadObject(file, obj, verbose = FALSE)
}
\arguments{
\item{file}{A RData file}

\item{obj}{Object name}

\item{verbose}{Whether the loading process should be verbose, see \code{\link{load}}}
}
\description{
Load an object by its name from a RData file
}

