% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFile.R
\name{loadObjectInEnv}
\alias{loadObjectInEnv}
\title{Load objects from a RData file and return them in an environment}
\usage{
loadObjectInEnv(file, obj = NULL, verbose = FALSE)
}
\arguments{
\item{file}{A RData file}

\item{obj}{Character string(s), optional object names. If set as \code{NULL}, all objects are returned}

\item{verbose}{Whether the loading process should be verbose, see \code{\link{load}}}
}
\value{
An environment containing the loaded objects.
}
\description{
Load objects from a RData file and return them in an environment
}
