% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exprs_matrix.R
\name{optional_suppress_warning}
\alias{optional_suppress_warning}
\title{Supress warning optionally}
\usage{
optional_suppress_warning(expr, suppress = TRUE)
}
\arguments{
\item{expr}{R expression}

\item{suppress}{Logical, whether or not to suppress warnings}
}
\value{
side effect is used
}
\description{
Supress warning optionally
}
