% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampliseq.R
\name{read_ampliseq_amplicons}
\alias{read_ampliseq_amplicons}
\title{Read AmpliSeq amplicon informaiton from a BED file}
\usage{
read_ampliseq_amplicons(bedFile)
}
\arguments{
\item{bedFile}{Character string, a BED file}
}
\value{
A \code{data.frame} containing following information: Amplicon, GeneSymbol, RefSeq, Length

The function is used to read Amplicon information from BED file
}
\description{
Read AmpliSeq amplicon informaiton from a BED file
}
