% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_david.R
\name{read_david}
\alias{read_david}
\title{Read tab-delimited result files from DAVID Bioinformatics Resources}
\usage{
read_david(file)
}
\arguments{
\item{file}{A file from DAVID Bioinformatics Resources}
}
\value{
A \code{data.frame} containing all information encoded in the file
}
\description{
Read tab-delimited result files from DAVID Bioinformatics Resources
}
\examples{
davidFile <- system.file("extdata/example-DAVID-output-subset.txt", package="ribiosIO")
davidResult <- read_david(davidFile)
}

