% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iofile.R
\name{setDataDir}
\alias{setDataDir}
\title{Set the data directory}
\usage{
setDataDir(path)
}
\arguments{
\item{path}{Path to the data directory}
}
\value{
\code{NULL}
The value is set in the options
}
\description{
Set the data directory
}
