% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampliseq.R
\name{read_ampliseq_bedcovgct}
\alias{read_ampliseq_bedcovgct}
\title{Read bedcov output of AmpliSeq amplicons and convert them to read counts}
\usage{
read_ampliseq_bedcovgct(file, bedFile)
}
\arguments{
\item{file}{Character string, a GCT file containing bedcov output of amplicons}

\item{bedFile}{Character string, a BED file encoding amplicons}
}
\value{
A \code{GctMatrix} object containing read counts

The function is used to convert read base counts returned by \code{samtools bedcov} to read counts using Amplicon information encoded in the bed file
}
\description{
Read bedcov output of AmpliSeq amplicons and convert them to read counts
}
