% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cls.R
\name{read_cls}
\alias{read_cls}
\title{Read cls files}
\usage{
read_cls(cls.file, offset = 0)
}
\arguments{
\item{cls.file}{CLS file name}

\item{offset}{The integer representing the first level, default is set to 0, for some software it can be set to 1

It reads CLS format files with space or tab as separators.}
}
\description{
Read cls files
}
\note{
The original CLS format specifies both tab or space can be used as separators.
This makes it unable to represent factors with sapces in levels. In order to accomodate
CLS format for these factors, we propose using tab as separators in CLS files when encoding factors
in R.

Because of this, only use \code{read_cls} if you are sure that no level string contain space(s).
Otherwise, use read_factor and write_factor instead: they can handle such strings
}
\examples{
idir <- system.file("extdata", package="ribiosIO")
sample.cls <- read_cls(file.path(idir, "test.cls"))
expFac <- factor(c("Case", "Control")[c(1,0,1,0,0,1,0,0,0,1,0,1,0,0,0,1,0,1,0,0,1,1,1,1,0,0)+1],
levels=c("Case", "Control"))
stopifnot(identical(sample.cls, expFac))
}
\seealso{
\code{\link{read_factor}} and \code{\link{write_factor}}
}

