% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosIO.R
\docType{package}
\name{ribiosIO-package}
\alias{ribiosIO-package}
\title{High-performance file input/output with ribios}
\description{
Provides high-performance file input/output for computational biology tasks
}
