% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosIO.R
\docType{package}
\name{ribiosIO-package}
\alias{ribiosIO}
\alias{ribiosIO-package}
\title{ribiosIO: Input/Output Utilities of the 'ribios' Suite}
\description{
Provides data structures and functions for file input/output in the 'ribios' software suite, supporting common bioinformatics and computational biology file formats, designed for fast loading and high performance with minimal dependencies.

Provides high-performance file input/output for bioinformatics and computational biology tasks
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bedapub/ribiosIO}
  \item Report bugs at \url{https://github.com/bedapub/ribiosIO/issues}
}

}
\author{
\strong{Maintainer}: Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID}) [contributor]

Other contributors:
\itemize{
  \item Balazs Banfai [contributor]
  \item F.Hoffmann-La Roche AG [copyright holder]
}

}
\keyword{internal}
